refactor :  도메인별로 폴더 나누기
